Add Tests for Description
Rework TestDisplayName for similarities with Description
